--- a/src/Tools/CsvImporter/Services/CsvImporterSqliteService/CsvImportSqliteServiceOptions.cs
+++ b/src/Tools/CsvImporter/Services/CsvImporterSqliteService/CsvImportSqliteServiceOptions.cs
@@ -1,0 +1,12 @@
+namespace EntraMfaPrefillinator.Tools.CsvImporter.Services;
+
+/// <summary>
+/// Options for configuring the <see cref="ICsvImporterSqliteService"/>.
+/// </summary>
+public class CsvImporterSqliteServiceOptions
+{
+    /// <summary>
+    /// The path to the SQLite database.
+    /// </summary>
+    public string DbPath { get; set; } = null!;
+}