This PR adds SQLite to `CsvImporter`. This change brings a massive bump in performance. Instead of reading the last run from a CSV file, entries from the CSV file are inserted/updated in a local SQLite database. 

For comparison, this is roughly what I've seen improved (With CSV files that have over 140,000 lines):

| Metric | Old | New |
| --- | --- | --- |
| **CPU utilization** | Roughly `100%` | Roughly `10-20%` |
| **Memory usage** | Roughly `200 MB` | Roughly `50-70 MB` |
| **Total runtime** | Roughly `40 seconds` | Roughly `2-30 seconds` |

A lot of these improvements are dependent on what it's ran on. A MacBook Pro with a M2 Pro was on the lower end of the improvements and was much faster; however, a Windows Server VM (Running on a `x86/x64` CPU) was on the higher end of the improvements. For the latter, it was still significantly faster than before and is overall more efficient. It just wasn't as fast on my dev machine.